requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cloudinary==1.29.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
